+++ src/sample.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-console */
-  import * as gerar from './index.js';
+
+import * as gerar from './index.js';
 
 /* ******************** DOCUMENTOS ********************** */
 // Gera CNH aleatória
@@ -66,6 +67,29 @@ console.log(gerar.tituloEleitor(false, 'PR'));
 
 // Gera Passaporte aleatório
 console.log(gerar.passaporte());
+
+// Gera uma Certidão (Nascimento, Casamento e Óbito)
+// Certidão aleatória
+console.log(gerar.certidao.aleatoria());       // com máscara
+console.log(gerar.certidao.aleatoria(false));  // sem máscara
+
+// Certidão de nascimento
+console.log(gerar.certidao.nascimento());      // com máscara
+console.log(gerar.certidao.nascimento(false)); // sem máscara
+
+// Certidão de casamento
+console.log(gerar.certidao.casamento());       // com máscara
+console.log(gerar.certidao.casamento(false));  // sem máscara
+
+// Certidão de óbito
+console.log(gerar.certidao.obito());           // com máscara
+console.log(gerar.certidao.obito(false));      // sem máscara
+
+// Objeto com os detalhes de Certidão
+console.log(gerar.certidao.aleatoriaObj());    // Objeto detalhado de certidão aleatória
+console.log(gerar.certidao.nascimentoObj());   // Objeto detalhado de certidão de nascimento
+console.log(gerar.certidao.casamentoObj());    // Objeto detalhado de certidão de casamento
+console.log(gerar.certidao.obitoObj());        // Objeto detalhado de certidão de óbito
 
 /* ******************** BANCOS ************************** */
 // Gera dados de um banco aleatório